+++ CHANGELOG.md
@@ -16,12 +16,16 @@
 ## v1.13.16 (NOT RELEASED YET)
 
 ### Added
+- `bit-cov-analyzer`
+  - progress updates while running
+  - zero-coverage region outputs now also generated
 - `bit-extract-seqs`
   - enables pulling out target seqs from a fasta by bed file or specified primers via subcommands
 - `--fragment-size-range` option added to `bit-gen-reads`, defaults to 10% of fragment size
 
 ### Changed
 - `bit-cov-analyzer`
+  - `-s | --sliding-window-size` changed to `-w | --window-size`, and `-S | --step-size` changed to `-s | --step-size` (lower-case)
   - drastic improvements to efficiency when working with large genomes (e.g., 3GB)
   - histogram of coverages no longer plotted by default, only done now when adding the `--write-window-stats` flag
 - `bit-get-mapped-reads-pid`
@@ -32,7 +36,9 @@
 - `bit-summarize-assembly`
   - adds commas when printing stats to terminal for readability
 - `bit-extract-seqs-by-coords` is now combined into `bit-extract-seqs`
-- `bit-gen-reads` now has a `--fragment-size-range` that defaults to 10% of fragment size
+- `bit-gen-reads`
+  - now has a `--fragment-size-range` that defaults to 10% of fragment size
+  - by default will not include regions with Ns in generated reads, add `--include-Ns` to allow that
 
 ---